refactor(cli): move options around
